pyproject.toml
[tool.poetry]
name = "book-review"
version = "0.1.0"
description = ""
authors = ["Luiz Lima <luizfelipevbll@gmail.com>"]
readme = "README.md"
packages = [{include = "book_review"}]

[tool.poetry.dependencies]
python = "^3.10"
django-ninja = "^0.19.1"
gunicorn = "^20.1.0"
Pillow = "^9.2.0"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
psycopg2 = "^2.9.3"
requests = "^2.28.1"
django-db-logger = "^0.1.12"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.21.0"
redis = "^4.3.4"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
ipython = "^8.6.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
